Fixed bug in dealing with repositories with untrusted hgrc files
By default, if there's an untrusted hgrc file, Mercurial prints a warning
about this. This warning was causing aHg to fail in parsing the output of
nearly all hg calls.

Fixed by making sure that the config entry "ui.report_untrusted" is set to
false when we need to parse the output of hg